Added new API call antiCacheEvictBlock() that will force the EE to evict a block of tuples out to the AntiCacheDB. Cleaned up PersistentTable::evictBlockToDisk() to use blockSize as the # of bytes to evict rather than the number of tuples

+++ src/ee/storage/persistenttable.h
@@ -247,7 +247,7 @@ class PersistentTable : public Table {
     // ANTI-CACHING OPERATIONS
     // ------------------------------------------------------------------
 #ifdef ANTICACHE
-    bool evictBlockToDisk(int block_size);
+    bool evictBlockToDisk(long block_size);
     TableTuple* createEvictedTuple(TableTuple &source_tuple, uint16_t block_id);
     bool readEvictedBlock(uint16_t block_id);
     bool mergeUnevictedTuples();
